Updated by songjie

+++ digiagent/src/main/java/net/cominfo/digiagent/service/CategoryService.java
@@ -4,8 +4,11 @@ import java.util.Date;
 import java.util.List;
 
 import net.cominfo.digiagent.persistence.dao.CategoryDao;
+import net.cominfo.digiagent.persistence.dao.ProductDao;
 import net.cominfo.digiagent.persistence.domain.Category;
 import net.cominfo.digiagent.persistence.domain.CategoryCriteria;
+import net.cominfo.digiagent.persistence.domain.Product;
+import net.cominfo.digiagent.persistence.domain.ProductCriteria;
 import net.cominfo.digiagent.persistence.domain.CategoryCriteria.Criteria;
 import net.cominfo.digiagent.persistence.sqlmapdao.PaginationContext;
 
@@ -20,6 +23,8 @@ public class CategoryService {
 
 	@Autowired
 	private CategoryDao categoryDao;
+	@Autowired
+	private ProductDao productDao;
 
 	public int countCategory() {
 		return categoryDao.countByExample(new CategoryCriteria());
@@ -67,8 +72,14 @@ public class CategoryService {
 		}
 	}
 	
-	public void delete(Integer id){
-		categoryDao.deleteByPrimaryKey(id);
+	public String delete(Integer id){
+		// 是否有产品关联
+		if (isReferenceProduct(id)) {
+			return "reference";
+		} else {
+			categoryDao.deleteByPrimaryKey(id);
+			return "success";
+		}
 	}
 	
 	private Category validateCategoryName(Category category) {
@@ -84,6 +95,18 @@ public class CategoryService {
 			return category;
 		} else {
 			return category;
+		}
+	}
+	
+	private boolean isReferenceProduct(Integer categoryId) {
+		ProductCriteria example = new ProductCriteria();
+		net.cominfo.digiagent.persistence.domain.ProductCriteria.Criteria criteria = example.createCriteria();
+		criteria.andCategoryIdEqualTo(categoryId);
+		List<Product> list = productDao.selectByExample(example);
+		if (list != null && list.size() > 0) {
+			return true;
+		} else {
+			return false;
 		}
 	}
 }
